app-history/per-entry-events/beforedispose-same-document-replaceState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(t => {
  const entriesBefore = appHistory.entries();
  const currentBefore = appHistory.current;
  const hrefBefore = location.href;

  let beforedisposeCalled = false;
  appHistory.current.onbeforedispose = t.step_func(e => {
    beforedisposeCalled = true;

    assert_equals(e.constructor, Event);
    assert_equals(e.bubbles, false);
    assert_equals(e.cancelable, false);
    assert_equals(e.composed, false);

    assert_array_equals(appHistory.entries(), entriesBefore);
    assert_equals(appHistory.current, currentBefore);
    assert_equals(appHistory.transition, null);
    assert_equals(location.href, hrefBefore);
  });

  history.replaceState(null, "", "?replacement");
  assert_true(beforedisposeCalled);
  assert_equals((new URL(location.href)).search, "?replacement");
  assert_not_equals(appHistory.current, currentBefore);
}, "beforedispose events when doing a same-document replace using history.replaceState()");
</script>
